segregacja i dodanie rozszerze?

+++ share/bslib/lock/script_unlock.sh
@@ -4,6 +4,6 @@ script_unlock(){
 	# mimo, że jest to jedno polecenie to zróbmy funkcję
 	# nieszczęściem by było usunąć niewłaściwy plik
 	# a tak minimalizujemy prawdomodobieństwo błędu
-	rm -f "$SCRIPT_LOCK"
-	echo -e "${B}[ zakonczono ]${N} usuwam plik blokujący skryptu: $SCRIPT_LOCK"
+	[[ -f $SCRIPT_LOCK ]] && rm -f "$SCRIPT_LOCK"
+	verbose "[ zakonczono ] usuwam plik blokujący skryptu: $SCRIPT_LOCK"
 }
